Enable typing per PEP-561
All that needs to be done is to add a "py.typed" file to the package,
since adequate type annotations exist in all of the dataclassy code.

Also addded __all__ to the __init__.py so that mypy can figure out what
is supposed to be exported. Without that we get missing type annotations
on import.