Extend differences table
Most users likely do not have a deep understanding of metaclasses. So
the error message "TypeError: metaclass conflict: the metaclass of a
derived class must be a (non-strict) subclass of the metaclasses of all
its bases"" will likely not be particularly helpful to them without some
head scratching. Better to include a note in the docs